Test5: improved time logs with offset checks for latency tests
