+++ client/components/Providers/ClientShell.tsx
@@ -1,5 +1,4 @@
 "use client";
-import { syncClock } from "@/lib/helpers";
 import { useEffect, useState } from "react";
 
 declare global {
@@ -14,15 +13,6 @@ export default function ClientShell({
   children: React.ReactNode;
 }) {
   const [mounted, setMounted] = useState(false);
-
-  useEffect(() => {
-    async function runSync() {
-      const offset = await syncClock();
-      window.swtf_offset = offset;
-      console.log("Clock offset synced:", offset);
-    }
-    runSync();
-  }, []);
 
   useEffect(() => {
     setMounted(true);
